Feature: Ability to start MotoServer within Python
