+++ moto/moto_server/templates/dashboard.html
